requirement.txt
requests2
bs4
pathlib